Revert "Calculate money made from assembly lines while closing"
This reverts commit 35c7b2649e575bfc297da13b55cde0f68fcafa4e.

+++ Assets/Scripts/ScriptableObjects/Factory_SO.cs
@@ -13,7 +13,7 @@ public class Factory_SO : ScriptableObject {
     public DateTime_SO LastTimeLogin;
 
     [Header("Factory Money")]
-    public double FactoryMoney= 2000;
+    public float FactoryMoney= 2000;
     public int companyLevel;
 
     [Header("Events")]
@@ -63,7 +63,7 @@ public class Factory_SO : ScriptableObject {
             Debug.Log("Factory doesn't have enough money.");
     }
 
-    public void DepositMoney(double money)
+    public void DepositMoney(float money)
     {
         FactoryMoney += money;
         factoryMoneyChanged.Raise();
